pc-1841 refacto and add snapshot
