+++ src/components/pages/Offerers/OffererItem/__specs__/PendingOffererItem.spec.jsx
@@ -1,0 +1,25 @@
+import React from 'react'
+import { shallow } from 'enzyme'
+
+import PendingOffererItem from '../PendingOffererItem'
+
+describe('src | components | pages | Offerers | OffererItem | PendingOffererItem', () => {
+  let props
+
+  beforeEach(() => {
+    props = {
+      offerer: {}
+    }
+  })
+
+  describe('snapshot', () => {
+    it('should match snapshot', () => {
+      // when
+      const wrapper = shallow(<PendingOffererItem {...props} />)
+
+      // then
+      expect(wrapper).toBeDefined()
+      expect(wrapper).toMatchSnapshot()
+    })
+  })
+})
